Refactor vehicle model usage and fix line endings
Changes made:
- Replaced usage of `Vehicles` model with `Vehicle` model throughout the application
- Fixed inconsistent line endings (converted from LP to CRLF)
- Corrected syntax errors in VehicleController (removed stray characters)
- Updated pagination and query methods to use proper model
- Cleaned up controller method formatting

Affected files:
- app/Http/Controllers/Controller.php
- app/Models/Vehicle.php
- app/Models/Vehicles.php
- database/migrations/...vehicles_table.php
- database/migrations/...reservations_table.php
- .../2025_04_01_083119_images.php
